Add RawPlayer and "Join Room" to the Edit page
